requirements.txt
django
django-filter
django-storages
pillow
gunicorn
whitenoise
dj-database-url